+++ interp.test/src/test/java/com/alexshtf/interp/test/ThreePointsCircleTest.java
@@ -35,8 +35,8 @@ public class ThreePointsCircleTest {
                                                  @LinSpace(first = 0, last = 4.99, count = 100) double radius,
                                                  @LinSpace(first = 0, last = 2 * Math.PI, count = 100) double angle) {
 
-        float x = (float)(1 + radius * Math.cos(angle));
-        float y = (float)(1 + radius * Math.sin(angle));
+        double x = 1 + radius * Math.cos(angle);
+        double y = 1 + radius * Math.sin(angle);
 
         assertTrue(getMessage(radius, angle, x, y), circle.isInside(new Point(x, y)));
     }
@@ -46,13 +46,13 @@ public class ThreePointsCircleTest {
                                                    @LinSpace(first = 5.01, last = 100, count = 100) double radius,
                                                    @LinSpace(first = 0, last = 2 * Math.PI, count = 100) double angle)
     {
-        float x = (float)(1 + radius * Math.cos(angle));
-        float y = (float)(1 + radius * Math.sin(angle));
+        double x = 1 + radius * Math.cos(angle);
+        double y = 1 + radius * Math.sin(angle);
 
         assertFalse(getMessage(radius, angle, x, y), circle.isInside(new Point(x, y)));
     }
 
-    private String getMessage(double radius, double angle, float x, float y) {
+    private String getMessage(double radius, double angle, double x, double y) {
         return String.format("x = %f, y = %f, radius = %f, angle = %f", x, y, radius, angle);
     }
 }
